--- google_api/Google/TODO.txt
@@ -1,18 +1,0 @@
-Gmail
-1x Send an automated email to nycdsa.student@gmail.com
-2x Send an automated email with attachment
-3. Schedule an automated email
-4. Read the contents of a received email
-
-GSheets
-1. Read contents of google spreadsheet to csv
-2. append data to google spreadsheet
-3. Create new google spreadsheet
-
-
-GDrive
-1. Upload file to google drive
-
-
-GForm
-1. Create new form from json structure.
