+++ google_api/credentials.json
@@ -1,0 +1,1 @@
+{"installed":{"client_id":"982051512742-g2ba1cti9cq3kb5pmisve5tcon8jjq4u.apps.googleusercontent.com","project_id":"quickstart-1566241139605","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://oauth2.googleapis.com/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"YcWUOqa11yE-0LR5wGOwtF8O","redirect_uris":["urn:ietf:wg:oauth:2.0:oob","http://localhost"]}}
